refactor!: drop naming series tool
