--- erpnext/setup/doctype/naming_series/README.md
@@ -1,1 +1,0 @@
-Tool to set numbering (naming) series for various DocTypes.
