--- erpnext/setup/doctype/naming_series/naming_series.json
@@ -1,132 +1,0 @@
-{
- "actions": [],
- "creation": "2022-05-26 03:12:49.087648",
- "description": "Set prefix for numbering series on your transactions",
- "doctype": "DocType",
- "engine": "InnoDB",
- "field_order": [
-  "setup_series",
-  "select_doc_for_series",
-  "help_html",
-  "naming_series_to_check",
-  "preview",
-  "add_series",
-  "set_options",
-  "user_must_always_select",
-  "update",
-  "column_break_13",
-  "update_series",
-  "prefix",
-  "current_value",
-  "update_series_start"
- ],
- "fields": [
-  {
-   "description": "Set prefix for numbering series on your transactions",
-   "fieldname": "setup_series",
-   "fieldtype": "Section Break",
-   "label": "Setup Series"
-  },
-  {
-   "fieldname": "select_doc_for_series",
-   "fieldtype": "Select",
-   "label": "Select Transaction"
-  },
-  {
-   "depends_on": "select_doc_for_series",
-   "fieldname": "help_html",
-   "fieldtype": "HTML",
-   "label": "Help HTML",
-   "options": "<div class=\"well\">\n    Edit list of Series in the box below. Rules:\n    <ul>\n        <li>Each Series Prefix on a new line.</li>\n        <li>Allowed special characters are \"/\" and \"-\"</li>\n        <li>\n            Optionally, set the number of digits in the series using dot (.)\n            followed by hashes (#). For example, \".####\" means that the series\n            will have four digits. Default is five digits.\n        </li>\n        <li>\n            You can also use variables in the series name by putting them\n            between (.) dots\n            <br>\n            Support Variables:\n            <ul>\n                <li><code>.YYYY.</code> - Year in 4 digits</li>\n                <li><code>.YY.</code> - Year in 2 digits</li>\n                <li><code>.MM.</code> - Month</li>\n                <li><code>.DD.</code> - Day of month</li>\n                <li><code>.WW.</code> - Week of the year</li>\n                <li><code>.FY.</code> - Fiscal Year</li>\n                <li>\n                    <code>.{fieldname}.</code> - fieldname on the document e.g.\n                    <code>branch</code>\n                </li>\n            </ul>\n        </li>\n    </ul>\n    Examples:\n    <ul>\n        <li>INV-</li>\n        <li>INV-10-</li>\n        <li>INVK-</li>\n        <li>INV-.YYYY.-.{branch}.-.MM.-.####</li>\n    </ul>\n</div>\n<br>\n"
-  },
-  {
-   "depends_on": "select_doc_for_series",
-   "fieldname": "set_options",
-   "fieldtype": "Text",
-   "label": "Series List for this Transaction"
-  },
-  {
-   "default": "0",
-   "depends_on": "select_doc_for_series",
-   "description": "Check this if you want to force the user to select a series before saving. There will be no default if you check this.",
-   "fieldname": "user_must_always_select",
-   "fieldtype": "Check",
-   "label": "User must always select"
-  },
-  {
-   "depends_on": "select_doc_for_series",
-   "fieldname": "update",
-   "fieldtype": "Button",
-   "label": "Update"
-  },
-  {
-   "description": "Change the starting / current sequence number of an existing series.",
-   "fieldname": "update_series",
-   "fieldtype": "Section Break",
-   "label": "Update Series"
-  },
-  {
-   "fieldname": "prefix",
-   "fieldtype": "Select",
-   "label": "Prefix"
-  },
-  {
-   "description": "This is the number of the last created transaction with this prefix",
-   "fieldname": "current_value",
-   "fieldtype": "Int",
-   "label": "Current Value"
-  },
-  {
-   "fieldname": "update_series_start",
-   "fieldtype": "Button",
-   "label": "Update Series Number",
-   "options": "update_series_start"
-  },
-  {
-   "fieldname": "naming_series_to_check",
-   "fieldtype": "Data",
-   "label": "Try a naming Series"
-  },
-  {
-   "default": " ",
-   "fieldname": "preview",
-   "fieldtype": "Text",
-   "label": "Preview of generated names",
-   "read_only": 1
-  },
-  {
-   "fieldname": "column_break_13",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "add_series",
-   "fieldtype": "Button",
-   "label": "Add this Series"
-  }
- ],
- "hide_toolbar": 1,
- "icon": "fa fa-sort-by-order",
- "idx": 1,
- "issingle": 1,
- "links": [],
- "modified": "2022-05-26 06:06:42.109504",
- "modified_by": "Administrator",
- "module": "Setup",
- "name": "Naming Series",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "email": 1,
-   "print": 1,
-   "read": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "read_only": 1,
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
